+++ src/main/java/com/followMe/delivery_server/delivery/application/DeliveryService.java
@@ -1,0 +1,9 @@
+package com.followMe.delivery_server.delivery.application;
+
+import com.followMe.delivery_server.delivery.application.dto.OrderCreateCommand;
+
+public interface DeliveryService {
+
+  void createDelivery(OrderCreateCommand command);
+
+}
