+++ src/main/java/com/followMe/delivery_server/delivery/application/DeliveryServiceImpl.java
@@ -1,11 +1,14 @@
-package com.followMe.delivery_server.delivery.service;
+package com.followMe.delivery_server.delivery.application;
 
-import com.followMe.delivery_server.client.HubClient;
+import com.followMe.delivery_server.delivery.infra.hub.HubClient;
 import com.followMe.delivery_server.delivery.domain.Delivery;
 import com.followMe.delivery_server.delivery.domain.Node;
-import com.followMe.delivery_server.delivery.dto.OrderCreateCommand;
-import com.followMe.delivery_server.delivery.repository.DeliveryRepository;
+import com.followMe.delivery_server.delivery.application.dto.DeliveryResponseDto;
+import com.followMe.delivery_server.delivery.application.dto.OrderCreateCommand;
+import com.followMe.delivery_server.delivery.infra.DeliveryQueryRepository;
+import com.followMe.delivery_server.delivery.domain.repository.DeliveryRepository;
 import java.util.List;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class DeliveryServiceImpl implements DeliveryService {
 
   private final DeliveryRepository deliveryRepository;
+  private final DeliveryQueryRepository deliveryQueryRepository;
   private final HubClient hubClient;
 
   @Transactional
