+++ src/main/java/com/followMe/delivery_server/delivery/domain/repository/DeliveryRepository.java
@@ -1,9 +1,11 @@
-package com.followMe.delivery_server.delivery.repository;
+package com.followMe.delivery_server.delivery.domain.repository;
 
 import com.followMe.delivery_server.delivery.domain.Delivery;
+import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryRepository extends JpaRepository<Delivery, UUID> {}
+public interface DeliveryRepository extends JpaRepository<Delivery, UUID> {
+}
