--- src/main/java/com/followMe/delivery_server/delivery/service/DeliveryService.java
@@ -1,8 +1,0 @@
-package com.followMe.delivery_server.delivery.service;
-
-import com.followMe.delivery_server.delivery.dto.OrderCreateCommand;
-
-public interface DeliveryService {
-
-  void createDelivery(OrderCreateCommand command);
-}
